[Pal/Linux-SGX] Make args_size/env_size size_t

+++ Pal/src/host/Linux-SGX/ecall_types.h
@@ -11,8 +11,8 @@ struct pal_sec;
 
 typedef struct {
     char * ms_args;
-    uint64_t ms_args_size;
+    size_t ms_args_size;
     char * ms_env;
-    uint64_t ms_env_size;
+    size_t ms_env_size;
     struct pal_sec * ms_sec_info;
 } ms_ecall_enclave_start_t;
